Fix level replay bug
While testing the Chrome builds, I noticed that pressing "Replay"
after completing level 1 resulted in an error. This is because
Game.start() was being called with the current level - 1, which
is 0 when on level 1.

The app was working without this code, which makes me think it
might be redundant; but I just fixed the bug for now.